--- src/__main__.py
@@ -1,5 +1,0 @@
-from src.app import app
-
-if __name__ == '__main__':
-    app.run(debug = True)
-
