+++ src/routes.py
@@ -1,38 +1,7 @@
-from flask import Flask, render_template, url_for, flash, redirect
-from flask_sqlalchemy import SQLAlchemy
+from flask import render_template, url_for, flash, redirect
+from src import app
 from src.forms import RegistrationForms, LoginForms
-from datetime import datetime
-
-import os
-
-app = Flask(__name__)
-app.config['SECRET_KEY'] = os.getenv('SECRET_KEY')
-app.config['SQLALCHEMY_DATABASE_URI'] = os.getenv('DATABASE_URI')
-app.config['SQLALCHEMY_TRACK_MODIFICATIONS'] = False
-db = SQLAlchemy(app)
-
-class User(db.Model):
-    id = db.Column(db.Integer, primary_key=True)
-    username = db.Column(db.String(20), unique=True, nullable=False)
-    email = db.Column(db.String(120), unique=True, nullable=False)
-    image_file = db.Column(db.String(20), nullable=False, default='default.jpeg')
-    password = db.Column(db.String(60), nullable=False)
-
-    posts = db.relationship('Post', backref='author', lazy=True)
-
-    def __repr__(self):
-        return f"User({self.username=}, {self.email=}, {self.image_file=})"
-
-class Post(db.Model):
-    id = db.Column(db.Integer, primary_key=True)
-    title = db.Column(db.String(100), nullable=False)
-    date_posted = db.Column(db.DateTime, nullable=False,
-        default=datetime.utcnow)
-    content = db.Column(db.Text, nullable=False)
-    user_id = db.Column(db.Integer, db.ForeignKey('user.id'), nullable=False)
-
-    def __repr__(self):
-        return f"Post({self.title=}, {self.date_posted=})"
+from src.models import User, Post
 
 posts = [
     {
@@ -76,5 +45,3 @@ def login():
             flash('Login Unsuccessful, Please check username and password',
                'danger')
     return render_template('login.html', title='Login', form=form)
-
-
